+++ xrpl/transaction/types/currency_amount.go
@@ -5,8 +5,6 @@ import (
 	"encoding/json"
 	"math/big"
 	"strconv"
-
-	"github.com/Peersyst/xrpl-go/xrpl/internal/flatten"
 )
 
 // CurrencyKind indicates the type of a currency amount (XRP, ISSUED, MPT).
@@ -89,7 +87,7 @@ func (i IssuedCurrencyAmount) Flatten() any {
 	json := make(map[string]any)
 
 	if i.Issuer != "" {
-		json["issuer"] = flatten.ClassicIfTagless(i.Issuer.String())
+		json["issuer"] = i.Issuer.Flatten()
 	}
 
 	if i.Currency != "" {