+++ xrpl/transaction/types/issued_currency.go
@@ -1,8 +1,6 @@
 //revive:disable:var-naming
 package types
 
-import "github.com/Peersyst/xrpl-go/xrpl/internal/flatten"
-
 // IssuedCurrency represents an amount of a non-XRP currency issued by an account.
 type IssuedCurrency struct {
 	Currency string  `json:"currency"`
@@ -13,6 +11,6 @@ type IssuedCurrency struct {
 func (i *IssuedCurrency) Flatten() map[string]any {
 	flattened := make(map[string]any)
 	flattened["currency"] = i.Currency
-	flattened["issuer"] = flatten.ClassicIfTagless(i.Issuer.String())
+	flattened["issuer"] = i.Issuer.Flatten()
 	return flattened
 }